Examples of combo and listpicker custom renderer
