+++ package.json
@@ -32,9 +32,9 @@
     "angular2-highlight-js": "^8.0.1",
     "hammerjs": "^2.0.8",
     "highlight.js": "^9.5.0",
-    "ontimize-web-ngx": "8.2.0",
-    "ontimize-web-ngx-theming": "8.0.1",
+    "ontimize-web-ngx": "file:ontimize-web-ngx-8.2.0.tgz",
     "ontimize-web-ngx-gallery": "8.0.1",
+    "ontimize-web-ngx-theming": "8.0.1",
     "rxjs": "~6.4.0",
     "tslib": "~1.10.0",
     "zone.js": "~0.9.1"
