+++ src/app/main/inputs/13.combo/input-combo.component.html
@@ -33,7 +33,8 @@
           fxFlex></o-combo>
         <o-combo attr="combo-editable-multiple-search" label="{{ 'INPUT.COMBOMULTIPLE.SEARCH' | oTranslate }}" [static-data]="getDataArray()"
           [data]="getValueMultiple()" value-column="key" columns="key;value" visible-columns="value" required="yes" read-only="no" multiple="yes"
-          null-selection="no" searchable="yes" fxFlex></o-combo>
+          null-selection="no" searchable="yes" fxFlex>
+        </o-combo>
       </div>
     </div>
     <div fxLayout="column" layout-padding>
