+++ src/app/main/inputs/16.listpicker-renderer/input-listpicker-custom-render.component.html
@@ -1,0 +1,11 @@
+<example-comp comp-name="INPUTS.LISTPICKER_CUSTOM_RENDERER" comp-desc="INPUTS.LISTPICKER_CUSTOM_RENDERER.DESC" [files]="files">
+  <o-form editable-detail="false" show-header="no" layout-direction="column">
+    <div fxLayout="column" layout-padding>
+      <label class="input-comp-title">{{ 'INPUT.BUTTON.LISTPICKER' | oTranslate }}</label>
+      <o-list-picker #listpicker attr="listpicker2" label="{{ 'INPUT.BUTTON.LISTPICKER_CUSTOM_RENDERER' | oTranslate }}" [static-data]="getDataArray()"
+        [data]="getValue()" filter="yes" value-column="key" columns="key;value" visible-columns="value" required="true" read-only="false">
+        <listpicker-custom-render></listpicker-custom-render>
+      </o-list-picker>
+    </div>
+  </o-form>
+</example-comp>
