+++ src/app/main/inputs/16.listpicker-renderer/listpicker-custom-render/listpicker-custom-render.component.html
@@ -1,0 +1,3 @@
+<ng-template #templateref let-value="value">
+  {{ getListPickerValue(value) }}
+</ng-template>
