+++ src/app/main/inputs/20.radio/input-radio.component.html
@@ -1,0 +1,24 @@
+<example-comp comp-name="INPUTS.RADIO" comp-desc="INPUTS.RADIO.DESC" [files]="files">
+  <o-form editable-detail="false" show-header="no">
+    <div fxLayout="column" layout-padding>
+      <label class="input-comp-title">{{ 'INPUTS.READ_ONLY' | oTranslate }}</label>
+      <o-radio attr="radio" label="{{ 'INPUT.BUTTON.RADIO' | oTranslate }}" [static-data]="getDataArray()" [data]="getValue()"
+        value-column="key" columns="key;value" visible-columns="value"></o-radio>
+    </div>
+    <div fxLayout="column" layout-padding>
+      <label class="input-comp-title">{{ 'INPUTS.EDITABLE' | oTranslate }}</label>
+      <o-row>
+        <o-radio attr="radio-editable" label="{{ 'INPUT.BUTTON.RADIO' | oTranslate }}" [static-data]="getDataArray()"
+          [data]="getValue()" value-column="key" columns="key;value" visible-columns="value" read-only="no" required="yes"></o-radio>
+        <o-radio attr="radio-editable2" label="{{ 'INPUT.BUTTON.RADIO' | oTranslate }}" [static-data]="getDataArray()"
+          [data]="getValue()" value-column="key" columns="key;value" visible-columns="value" read-only="no" required="yes"
+          layout="row"></o-radio>
+      </o-row>
+    </div>
+    <div fxLayout="column" layout-padding>
+      <label class="input-comp-title">{{ 'INPUTS.DISABLED' | oTranslate }}</label>
+      <o-radio attr="radio-disabled" label="{{ 'INPUT.BUTTON.RADIO' | oTranslate }}" [static-data]="getDataArray()"
+        [data]="getValue()" value-column="key" columns="key;value" visible-columns="value" enabled="no"></o-radio>
+    </div>
+  </o-form>
+</example-comp>
