--- src/app/main/inputs/21.daterange/date-range/date-range.component.html
@@ -1,53 +1,0 @@
-<div fxLayout="row" fxLayoutAlign="end center" class="selector-mode">
-  <mat-icon>desktop_mac</mat-icon>{{'MODE.DESKTOP' | oTranslate}}
-  <mat-slide-toggle #mode>
-    <mat-icon>smartphone</mat-icon> {{'MODE.MOBILE'| oTranslate}}
-  </mat-slide-toggle>
-</div>
-
-<example-comp #ex1 comp-name="INPUTS.DATERANGE" comp-desc="INPUTS.DATERAGE.DESC" [files]="files" (showSource)="onShowSource('o-daterange', ex1)">
-  <o-form editable-detail="no" show-header="no" layout-direction="column">
-    <div fxLayout="column" layout-padding>
-      <label class="input-comp-title">{{ 'INPUTS.READ_ONLY' | oTranslate }}</label>
-      <o-daterange-input attr="daterange1" label="{{ 'INPUT.BUTTON.DATERANGE' | oTranslate }}" required="yes" [data]="getValue()">
-      </o-daterange-input>
-    </div>
-    <div fxLayout="column" layout-padding>
-      <o-daterange-input attr="daterange2" label="{{ 'INPUT.BUTTON.DATERANGE' | oTranslate }}" read-only="no" required="yes" [data]="getValue()"
-        clear-button="yes" format="LL" separator=" to " [mode]="mode.checked?'mobile':'desktop'">
-      </o-daterange-input>
-    </div>
-    <div fxLayout="column" layout-padding>
-      <label class="input-comp-title">{{ 'INPUTS.DISABLED' | oTranslate }}</label>
-      <o-daterange-input attr="daterange3" label="{{ 'INPUT.BUTTON.DATERANGE' | oTranslate }}" enabled="no" [data]="getValue()">
-      </o-daterange-input>
-    </div>
-  </o-form>
-</example-comp>
-
-<example-comp #ex2 comp-name="INPUTS.DATERANGE.MINMAX" comp-desc="INPUTS.DATERANGE.MINMAX.DESC" [files]="mixMaxDateRangeFiles"
-  (showSource)="onShowSource('o-daterange-minmax', ex2)">
-
-  <o-form editable-detail="no" show-header="no" layout-direction="column">
-    <div fxLayout="column" layout-padding>
-      <label class="input-comp-title">{{ 'INPUTS.EDITABLE' | oTranslate }}</label>
-      <o-daterange-input attr="daterange4" label="{{ 'INPUT.BUTTON.DATERANGE' | oTranslate }}" read-only="no" required="yes" [data]="getValue()"
-        min="15/05/2019" max="20/06/2019" format="DD/MM/YYYY" show-week-numbers="true" [mode]="mode.checked?'mobile':'desktop'">
-      </o-daterange-input>
-    </div>
-
-  </o-form>
-</example-comp>
-<example-comp #ex3 comp-name="INPUTS.DATERANGE.SHOWRANGES" comp-desc="INPUTS.DATERANGE.SHOWRANGES.DESC" [files]="showRangeDateRangeFiles"
-  (showSource)="onShowSource('o-daterange-ranges', ex3)">
-
-  <o-form editable-detail="no" show-header="no" layout-direction="column">
-    <div fxLayout="column" layout-padding>
-      <label class="input-comp-title">{{ 'INPUTS.EDITABLE' | oTranslate }}</label>
-      <o-daterange-input attr="daterange5" label="{{ 'INPUT.BUTTON.DATERANGE' | oTranslate }}" read-only="no" required="yes" [data]="getValue()"
-        show-ranges="true" format="DD/MM/YYYY" separator=" to " touch-ui="yes" [mode]="mode.checked?'mobile':'desktop'">
-      </o-daterange-input>
-    </div>
-
-  </o-form>
-</example-comp>
