--- src/app/main/inputs/21.daterange/date-range/date-range.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { InputDateRangeComponent } from './date-range.component';
-
-
-describe('InputDateRangeComponent', () => {
-  let component: InputDateRangeComponent;
-  let fixture: ComponentFixture<InputDateRangeComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ InputDateRangeComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(InputDateRangeComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
