+++ src/app/main/inputs/23.daterange/date-range/date-range.component.scss
@@ -1,0 +1,16 @@
+.selector-mode {
+  .mat-slide-toggle {
+    margin-right: 8px;
+    margin-left: 8px;
+  }
+
+  .mat-icon {
+    padding-right: 8px;
+    opacity: .8;
+  }
+
+  .mat-slide-toggle-content {
+    display: flex;
+
+  }
+}
