+++ src/app/main/inputs/inputs.component.html
@@ -1,6 +1,7 @@
 <ul class="list-grid-container">
   <li class="list-grid-button"><a routerLink="/main/inputs/checkbox" routerLinkActive="active" fxFlex>&#8226; {{ 'INPUT.BUTTON.CHECKBOX' | oTranslate }}</a></li>
   <li class="list-grid-button"><a routerLink="/main/inputs/combo" routerLinkActive="active" fxFlex>&#8226; {{ 'INPUT.BUTTON.COMBO' | oTranslate }}</a></li>
+  <li class="list-grid-button"><a routerLink="/main/inputs/combo-custom-renderer" routerLinkActive="active" fxFlex>&#8226; {{ 'INPUT.BUTTON.COMBO_CUSTOM_RENDERER' | oTranslate }} <span class="new">new</span></a></li>
   <li class="list-grid-button"><a routerLink="/main/inputs/currency" routerLinkActive="active" fxFlex>&#8226; {{ 'INPUT.BUTTON.CURRENCY' | oTranslate }}</a></li>
   <li class="list-grid-button"><a routerLink="/main/inputs/date" routerLinkActive="active" fxFlex>&#8226; {{ 'INPUT.BUTTON.DATE' | oTranslate }}</a></li>
   <li class="list-grid-button"><a routerLink="/main/inputs/daterange" routerLinkActive="active" fxFlex>&#8226; {{ 'INPUT.BUTTON.DATERANGE' | oTranslate }}</a></li>
@@ -10,6 +11,7 @@
   <li class="list-grid-button"><a routerLink="/main/inputs/html" routerLinkActive="active" fxFlex>&#8226; {{ 'INPUT.BUTTON.HTML' | oTranslate }}</a></li>
   <li class="list-grid-button"><a routerLink="/main/inputs/integer" routerLinkActive="active" fxFlex>&#8226; {{ 'INPUT.BUTTON.INTEGER' | oTranslate }}</a></li>
   <li class="list-grid-button"><a routerLink="/main/inputs/listpicker" routerLinkActive="active" fxFlex>&#8226; {{ 'INPUT.BUTTON.LISTPICKER' | oTranslate }}</a></li>
+  <li class="list-grid-button"><a routerLink="/main/inputs/listpicker-custom-renderer" routerLinkActive="active" fxFlex>&#8226; {{ 'INPUT.BUTTON.LISTPICKER_CUSTOM_RENDERER' | oTranslate }} <span class="new">new</span></a></li>
   <li class="list-grid-button"><a routerLink="/main/inputs/nif" routerLinkActive="active" fxFlex>&#8226; {{ 'INPUT.BUTTON.NIF' | oTranslate }}</a></li>
   <li class="list-grid-button"><a routerLink="/main/inputs/password" routerLinkActive="active" fxFlex>&#8226; {{ 'INPUT.BUTTON.PASSWORD' | oTranslate }}</a></li>
   <li class="list-grid-button"><a routerLink="/main/inputs/percent" routerLinkActive="active" fxFlex>&#8226; {{ 'INPUT.BUTTON.PERCENT' | oTranslate }}</a></li>
