+++ src/app/main/main.module.ts
@@ -1,8 +1,9 @@
 import { NgModule } from '@angular/core';
 import { OntimizeWebModule } from 'ontimize-web-ngx';
+
 import { SharedModule } from '../shared/shared.module';
-import { MainComponent } from './main.component';
 import { MainRoutingModule } from './main-routing.module';
+import { MainComponent } from './main.component';
 
 @NgModule({
   imports: [
