+++ src/assets/i18n/en.json
@@ -144,8 +144,10 @@
   "INPUT.BUTTON.FILE": "File",
   "INPUT.BUTTON.CHECKBOX": "Checkbox",
   "INPUT.BUTTON.LISTPICKER": "List picker",
+  "INPUT.BUTTON.LISTPICKER_CUSTOM_RENDERER": "List picker Custom Renderer",
   "INPUT.BUTTON.VALIDATORS": "Validators",
   "INPUT.BUTTON.COMBO": "Combo",
+  "INPUT.BUTTON.COMBO_CUSTOM_RENDERER": "Combo Custom Renderer",
   "INPUT.BUTTON.HOUR": "Hour",
   "INPUT.BUTTON.TIME": "Time",
   "INPUT.BUTTON.RADIO": "Radio",
@@ -156,8 +158,10 @@
   "INPUTS.SLIDER.DESC": "A component allows for the selection of a value from a range.",
   "INPUT.COMBO": "Combo",
   "INPUT.COMBO.SEARCH": "Combo with search",
+  "INPUT.COMBO_CUSTOM_RENDERER.SEARCH": "Combo with search and custom renderer",
   "INPUT.COMBOMULTIPLE": "Combo multiple",
   "INPUT.COMBOMULTIPLE.SEARCH": "Combo multiple with search",
+  "INPUT.COMBOMULTIPLE_CUSTOM_RENDERER.SEARCH": "Combo multiple with search and custom renderer",
   "INPUTS.CLEAR": "Clear button",
   "INPUTS.CLEAR.DESC": "Add a button for clearing the input field using the attribute 'clear-button'",
   "INPUTS.TEXT": "Text inputs",
@@ -198,11 +202,17 @@
   "INPUTS.CHECKBOXBOOLEANTYPE_BOOLEAN": "Boolean",
   "INPUTS.CHECKBOXBOOLEANTYPE.DESC": "There are 3 boolean types (number, string and boolean) ",
   "INPUTS.COMBO": "Combo inputs",
+  "INPUTS.COMBO_CUSTOM_RENDERER": "Combo inputs custom renderer",
   "INPUTS.COMBO_MULTIPLE": "Combo inputs (multiple selection)",
+  "INPUTS.COMBO_MULTIPLE_CUSTOM_RENDERER": "Combo inputs (multiple selection) with custom renderer",
   "INPUTS.COMBO.DESC": "A component for getting or displaying an option from multiple.",
+  "INPUTS.COMBO_CUSTOM_RENDERER.DESC": "A component for getting or displaying an option from multiple with custom renderer.",
   "INPUTS.COMBO.DESC_MULTIPLE": "A combo component that allows multiple selection",
+  "INPUTS.COMBO_CUSTOM_RENDERER.DESC_MULTIPLE": "A combo component that allows multiple selection with custom renderer",
   "INPUTS.LISTPICKER": "List picker inputs",
+  "INPUTS.LISTPICKER_CUSTOM_RENDERER": "List picker custom renderer",
   "INPUTS.LISTPICKER.DESC": "A component for getting or displaying an option from multiple.",
+  "INPUTS.LISTPICKER_CUSTOM_RENDERER.DESC": "A component for getting or displaying an option from multiple with your own renderer.",
   "INPUTS.HOUR": "Hour",
   "INPUTS.HOUR.DESC": "A component for getting or displaying hour submitted by the user.",
   "INPUTS.HOUR.FORMATS": "Formats",
