+++ src/assets/i18n/es.json
@@ -140,8 +140,10 @@
   "INPUT.BUTTON.FILE": "Fichero",
   "INPUT.BUTTON.CHECKBOX": "Checkbox",
   "INPUT.BUTTON.LISTPICKER": "List picker",
+  "INPUT.BUTTON.LISTPICKER_CUSTOM_RENDERER": "List picker - Renderer personalizado",
   "INPUT.BUTTON.VALIDATORS": "Validadores",
   "INPUT.BUTTON.COMBO": "Combo",
+  "INPUT.BUTTON.COMBO_CUSTOM_RENDERER": "Combo - Renderer personalizado",
   "INPUT.BUTTON.HOUR": "Hora",
   "INPUT.BUTTON.TIME": "Fecha y hora",
   "INPUT.BUTTON.RADIO": "Radio",
@@ -150,8 +152,10 @@
   "INPUT.BUTTON.DATERANGE": "Rango de fechas",
   "INPUT.COMBO": "Combo",
   "INPUT.COMBO.SEARCH": "Combo con búsqueda",
+  "INPUT.COMBO_CUSTOM_RENDERER.SEARCH": "Combo con búsqueda y renderer personalizado",
   "INPUT.COMBOMULTIPLE": "Combo multiple",
   "INPUT.COMBOMULTIPLE.SEARCH": "Combo multiple con búsqueda",
+  "INPUT.COMBOMULTIPLE_CUSTOM_RENDERER.SEARCH": "Combo multiple con búsqueda y renderer personalizado",
   "INPUTS.CLEAR": "Botón limpiar",
   "INPUTS.CLEAR.DESC": "Añade un botón para limpiar el campo usando el atributo 'clear-button'",
   "INPUTS.TEXT": "Inputs de texto",
@@ -192,11 +196,17 @@
   "INPUTS.CHECKBOXBOOLEANTYPE_STRING": "String",
   "INPUTS.CHECKBOXBOOLEANTYPE_BOOLEAN": "Boolean",
   "INPUTS.COMBO": "Inputs combo",
+  "INPUTS.COMBO_CUSTOM_RENDERER": "Inputs combo renderer personalizado",
   "INPUTS.COMBO_MULTIPLE": "Combo inputs (multiple selection)",
+  "INPUTS.COMBO_MULTIPLE_CUSTOM_RENDERER": "Combo inputs (multiple selection) con renderer personalizado",
   "INPUTS.COMBO.DESC": "Componente para insertar o mostrar una opción entre múltiples.",
+  "INPUTS.COMBO_CUSTOM_RENDERER.DESC": "Componente para insertar o mostrar una opción entre múltiples con renderer personalizado.",
   "INPUTS.COMBO.DESC_MULTIPLE": "Componente que permite selección múltiple",
+  "INPUTS.COMBO_CUSTOM_RENDERER.DESC_MULTIPLE": "Componente que permite selección múltiple con renderer personalizado",
   "INPUTS.LISTPICKER": "Inputs list picker",
+  "INPUTS.LISTPICKER_CUSTOM_RENDERER": "List picker renderer personalizado",
   "INPUTS.LISTPICKER.DESC": "Componente para insertar o mostrar una opción entre múltiples.",
+  "INPUTS.LISTPICKER_CUSTOM_RENDERER.DESC": "Componente para insertar o mostrar una opción entre múltiples con renderer personalizado.",
   "INPUTS.HOUR": "Hora",
   "INPUTS.HOUR.DESC": "Componente para insertar o mostrar la hora.",
   "INPUTS.HOUR.FORMATS": "Formatos",
